+++ app/Plugin/Uninstaller.php
@@ -10,9 +10,11 @@ namespace PersonioIntegrationLight\Plugin;
 // prevent direct access.
 defined( 'ABSPATH' ) || exit;
 
+use PersonioIntegrationLight\Dependencies\easyTransientsForWordPress\Transient;
+use PersonioIntegrationLight\Dependencies\easyTransientsForWordPress\Transients;
 use PersonioIntegrationLight\Helper;
 use PersonioIntegrationLight\PersonioIntegration\Extensions;
-use PersonioIntegrationLight\PersonioIntegration\Imports;
+use PersonioIntegrationLight\PersonioIntegration\Personio_Accounts;
 use PersonioIntegrationLight\PersonioIntegration\Post_Type;
 use PersonioIntegrationLight\PersonioIntegration\Post_Types;
 use PersonioIntegrationLight\PersonioIntegration\PostTypes\PersonioPosition;
@@ -31,7 +33,7 @@ class Uninstaller {
 	private static ?Uninstaller $instance = null;
 
 	/**
-	 * Constructor for Init-Handler.
+	 * Constructor for this object.
 	 */
 	private function __construct() {}
 
@@ -46,11 +48,11 @@ class Uninstaller {
 	 * Return the instance of this Singleton object.
 	 */
 	public static function get_instance(): Uninstaller {
-		if ( ! static::$instance instanceof static ) {
-			static::$instance = new static();
+		if ( is_null( self::$instance ) ) {
+			self::$instance = new self();
 		}
 
-		return static::$instance;
+		return self::$instance;
 	}
 
 	/**
@@ -58,7 +60,7 @@ class Uninstaller {
 	 *
 	 * Either via uninstall or via cli.
 	 *
-	 * @param array $delete_data Marker to delete all data.
+	 * @param array<int> $delete_data Marker to delete all data.
 	 * @return void
 	 */
 	public function run( array $delete_data = array() ): void {
@@ -70,31 +72,103 @@ class Uninstaller {
 			$original_blog_id = get_current_blog_id();
 
 			// loop through the blogs.
-			foreach ( Helper::get_blogs() as $blog_id ) {
+			foreach ( Helper::get_blogs() as $blog ) {
 				// switch to the blog.
-				switch_to_blog( $blog_id->blog_id );
+				switch_to_blog( $blog->blog_id );
 
 				// run tasks for deactivation in this single blog.
-				$this->deactivation_tasks( $delete_data );
+				$this->deinstallation_tasks( $delete_data );
 			}
 
 			// switch back to original blog.
 			switch_to_blog( $original_blog_id );
 		} else {
 			// simply run the tasks on single-site-install.
-			$this->deactivation_tasks( $delete_data );
+			$this->deinstallation_tasks( $delete_data );
 		}
 	}
 
 	/**
 	 * Define the tasks to run during deactivation.
 	 *
-	 * @param array $delete_data Whether all data should be removed or not (should be an array with value 1 for yes).
+	 * @param array<int> $delete_data Whether all data should be removed or not (should be an array with value 1 for yes).
 	 *
 	 * @return void
 	 */
-	private function deactivation_tasks( array $delete_data ): void {
+	private function deinstallation_tasks( array $delete_data ): void {
 		global $wpdb;
+
+		// delete all plugin-data.
+		if ( ! empty( $delete_data[0] ) && 1 === absint( $delete_data[0] ) ) {
+			/**
+			 * Set all settings.
+			 */
+
+			// initialize the plugin.
+			Init::get_instance()->init();
+
+			/**
+			 * Run the global init to initialize all components.
+			 */
+			do_action( 'init' );
+
+			// enable the settings.
+			\PersonioIntegrationLight\Dependencies\easySettingsForWordPress\Settings::get_instance()->activation();
+
+			/**
+			 * Remove them.
+			 */
+
+			// clean managed settings.
+			\PersonioIntegrationLight\Dependencies\easySettingsForWordPress\Settings::get_instance()->delete_settings();
+
+			// initialize the extensions to call their uninstall routines later.
+			Extensions::get_instance()->init();
+
+			// reset Personio- and language-specific settings.
+			Personio_Accounts::get_instance()->reset_personio_settings();
+
+			// delete taxonomies.
+			Taxonomies::get_instance()->delete_all();
+
+			// delete position.
+			PersonioPosition::get_instance()->delete_positions();
+
+			// remove custom options.
+			foreach ( $this->get_options() as $option ) {
+				delete_option( $option );
+			}
+
+			// remove user meta for each cpt we provide.
+			foreach ( Post_Types::get_instance()->get_post_types() as $post_type ) {
+				// create classname.
+				$classname = $post_type . '::get_instance';
+
+				// bail if classname is not callable.
+				if ( ! is_callable( $classname ) ) {
+					continue;
+				}
+
+				// get the object.
+				$obj = $classname();
+
+				// bail if object is not our own post_type.
+				if ( ! $obj instanceof Post_Type ) {
+					continue;
+				}
+
+				// bail if post type is not from this plugin.
+				if ( ! $obj->is_from_plugin( WP_PERSONIO_INTEGRATION_PLUGIN ) ) {
+					continue;
+				}
+
+				// delete the settings of this object from user meta.
+				$wpdb->delete( $wpdb->usermeta, array( 'meta_key' => $obj->get_name() ) ); // phpcs:ignore WordPress.DB.DirectDatabaseQuery
+			}
+
+			// uninstall extensions.
+			Extensions::get_instance()->uninstall_all();
+		}
 
 		// remove schedules.
 		Schedules::get_instance()->delete_all();
@@ -104,6 +178,12 @@ class Uninstaller {
 
 		// remove transients.
 		foreach ( Transients::get_instance()->get_transients() as $transient_obj ) {
+			// bail if object is not ours.
+			if ( ! $transient_obj instanceof Transient ) { // @phpstan-ignore instanceof.alwaysTrue
+				continue;
+			}
+
+			// delete it.
 			$transient_obj->delete();
 			$transient_obj->delete_dismiss();
 		}
@@ -111,60 +191,23 @@ class Uninstaller {
 		// remove plugin update transient.
 		delete_transient( 'personio_integration_light_plugin_update_notices' );
 
-		// delete all plugin-data.
-		if ( ! empty( $delete_data[0] ) && 1 === absint( $delete_data[0] ) ) {
-			// initialize the extensions to call their uninstall routines later.
-			Extensions::get_instance()->init();
-
-			// reset Personio- and language-specific settings.
-			Imports::get_instance()->reset_personio_settings();
-
-			// delete taxonomies.
-			Taxonomies::get_instance()->delete_all();
-
-			// delete position.
-			PersonioPosition::get_instance()->delete_positions();
-
-			// remove options from settings.
-			$settings_obj = Settings::get_instance();
-			$settings_obj->set_settings();
-			foreach ( $settings_obj->get_settings() as $section_settings ) {
-				foreach ( $section_settings['fields'] as $field_name => $field_settings ) {
-					delete_option( $field_name );
-				}
-			}
-
-			// remove custom options.
-			foreach ( $this->get_options() as $option ) {
-				delete_option( $option );
-			}
-
-			// remove setup-options.
-			Setup::get_instance()->uninstall();
-
-			// remove user meta for each cpt we provide.
-			foreach ( Post_Types::get_instance()->get_post_types() as $post_type ) {
-				$obj = call_user_func( $post_type . '::get_instance' );
-				if ( $obj instanceof Post_Type && $obj->is_from_plugin( WP_PERSONIO_INTEGRATION_PLUGIN ) ) {
-					$wpdb->delete( $wpdb->usermeta, array( 'meta_key' => $obj->get_name() ) ); // phpcs:ignore WordPress.DB.DirectDatabaseQuery
-				}
-			}
-
-			// uninstall extensions.
-			Extensions::get_instance()->uninstall_all();
-		}
+		// remove setup-options.
+		Setup::get_instance()->uninstall();
 
 		// remove roles from our plugin.
 		Roles::get_instance()->uninstall();
 
 		// delete our custom database-tables.
 		Init::get_instance()->delete_db_tables();
+
+		// remove crypt settings.
+		Crypt::get_instance()->uninstall();
 	}
 
 	/**
 	 * Return list of options this plugin is using which are not configured via @file Settings.php.
 	 *
-	 * @return array
+	 * @return array<string>
 	 */
 	private function get_options(): array {
 		return array(
@@ -176,11 +219,13 @@ class Uninstaller {
 			WP_PERSONIO_INTEGRATION_DELETE_RUNNING,
 			WP_PERSONIO_INTEGRATION_DELETE_STATUS,
 			WP_PERSONIO_INTEGRATION_TRANSIENTS_LIST,
+			WP_PERSONIO_INTEGRATION_IMPORT_NEW_POSITIONS,
+			WP_PERSONIO_INTEGRATION_DELETE_COUNT,
+			WP_PERSONIO_INTEGRATION_DELETE_MAX,
 			'personioIntegrationLightInstallDate',
 			'personio_integration_settings',
-			'personio_integration_schedules',
-			'personioIntegrationPositionCount',
 			'personio_integration_intro',
+			'personioIntegrationPageBuilder',
 		);
 	}
 }
